+++ routers/router.go
@@ -20,7 +20,7 @@ func init() {
 	// Routes static pour protéger les photos des albums
 	nsStatic := web.NewNamespace("/s",
 		web.NSBefore(controllers.AuthRequiredProfile),
-		web.NSRouter("/image/:beefileid", &controllers.MainController{}, "get:Image"),
+		web.NSRouter("/document/:beefileid", &controllers.MainController{}, "get:Document"),
 	)
 
 	// Routes avec rôle reader